Initial Implementation of "occurrence-based" movement
Added new commands that do contextual movement based of of the starting cursor position. If it is inside of a scope and on a TokenBaseKind_Identifier it will move between occurrences of that identifier in the scope (outer most). Otherwise it will act the same as move_[up/down]_to_blank_line_end.